Enemy AI 행동 추가 및 UI 반응형 개선
적 AI에 Charge, Fire, Kiting 행동을 추가하고 개선했습니다. 주요 UI 위젯의 폰트/크기/위치가 해상도에 따라 자동 조정되도록 리팩토링했습니다. 신규 적(Spitter, Bruiser) 데이터가 추가되었고, 테스트용 적 스폰 설정이 변경되었습니다.

+++ DirectX/Game_2D/Character/CEnemy.h
@@ -72,7 +72,7 @@ protected:
 
 	bool bOnKiting{};
 
-	static constexpr int FIRE_COOLDOWN_TIME = 5.f;
+	static constexpr float FIRE_COOLDOWN_TIME = 5.f;
 	float ElapsedFromFire = std::numeric_limits<float>::infinity();
 
 	static constexpr float SPAWN_FLICKER_INTERVAL = 0.4f;
